requirements.txt
Flask
Pillow==10.0.1
selectivesearch
tensorflow


pyheif
imagehash
exifread
scikit-learn
torch
torchvision
pandas==1.5.3
opencv-python
PyPDF2
werkzeug
flask_cors